+++ CHANGES.md
@@ -17,6 +17,7 @@ unreleased
     - Add reproduction case for #1763 but it is not failing anymore (#2021)
     - Add a test to reproduce [Locate command fails on multi-line type definitions](#1987) (#2020)
     - Add a regression test for issue #2019 (#2030)
+    - Add a test reproducing the locate command failure when editor position is in another declaration (#2033)
 
 merlin 5.6.1
 Sat Dec 20 11:15:42 CET 2025